refactor cd
