+++ src/main/java/pl/sda/library/domain/service/CategoryService.java
@@ -3,7 +3,7 @@ package pl.sda.library.domain.service;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import pl.sda.library.domain.model.Category;
-import pl.sda.library.repository.CategoryRepository;
+import pl.sda.library.domain.port.CategoryRepository;
 import pl.sda.library.domain.model.exception.ObjectAlreadyExistException;
 import pl.sda.library.domain.model.exception.ObjectDoesNotExistException;
