+++ src/main/java/pl/sda/library/infrastructure/jpa/BookDtoRepository.java
@@ -1,6 +1,8 @@
-package pl.sda.library.repository;
+package pl.sda.library.infrastructure.jpa;
 
 import org.springframework.data.repository.CrudRepository;
+import pl.sda.library.infrastructure.jpa.BookDto;
+import pl.sda.library.repository.UserDto;
 
 interface BookDtoRepository extends CrudRepository<BookDto, Long> {
