+++ src/main/java/pl/sda/library/infrastructure/jpa/BookstoreDtoRepository.java
@@ -1,6 +1,8 @@
-package pl.sda.library.repository;
+package pl.sda.library.infrastructure.jpa;
 
 import org.springframework.data.repository.CrudRepository;
+import pl.sda.library.infrastructure.jpa.BookstoreDto;
+
 import java.util.Optional;
 
 interface BookstoreDtoRepository extends CrudRepository<BookstoreDto, Long> {
