+++ src/main/java/pl/sda/library/infrastructure/jpa/SeriesDtoRepository.java
@@ -1,6 +1,7 @@
-package pl.sda.library.repository;
+package pl.sda.library.infrastructure.jpa;
 
 import org.springframework.data.repository.CrudRepository;
+import pl.sda.library.infrastructure.jpa.SeriesDto;
 
 import java.util.Optional;
