+++ src/test/java/pl/sda/library/repository/DtoFactoryTest.java
@@ -7,6 +7,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import pl.sda.library.LibraryApplication;
 import pl.sda.library.domain.model.*;
+import pl.sda.library.infrastructure.jpa.BookDto;
+import pl.sda.library.infrastructure.jpa.DtoFactory;
+import pl.sda.library.infrastructure.jpa.UserDto;
 
 import java.time.LocalDate;
